BREAKING CHANGE(BTable): Updated the items processing logic, Now items are fully reactive and can be updated directly through the item's cell slot.
